requirements.txt
django>=1.11,<1.12
django-debug-toolbar==1.8
gunicorn==19.4.5
psycopg2==2.7.3.1
whitenoise==3.3.1
python-adjutant
